move 'SetupAssistant' out of 'pact' package - it's used in other integration tests
