+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/integration/pact/ReferralContracts.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.integration.pact
 
 import au.com.dius.pact.provider.junitsupport.State
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.integration.SetupAssistant
 import java.time.OffsetDateTime
 import java.util.UUID
