+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/integration/pact/ServiceCategoryContracts.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.integration.pact
 
 import au.com.dius.pact.provider.junitsupport.State
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.integration.SetupAssistant
 
 class ServiceCategoryContracts(private val setupAssistant: SetupAssistant) {
   @State("a service category with ID 428ee70f-3001-4399-95a6-ad25eaaede16 exists")
